Added Hashtag Indexing
The hashtag index is stored in a new variable. Storing the hashtag
field can be disabled by toggling the instance variable
useHashtagScoring.

Queries are compared to both the regular index and the hashtag index, a
sum of these two values is used as the overall score for that query.
Currently hashtag documents have a scoring boost boost of 1.00,
alternative values could be used/tested.

The flow for measuring similarity of queries in the data is

Measure TFIDF on full tweets

Measure TFIDF on hashtag index

Combine top results from both sets by summing scores across by
similarity measures

Choose the top 1000 (or all available) results from this combined set,